调整格式、名称以及部分debug

+++ catHospital/src/main/java/com/renkaen/cat_hospital/service/impl/security/LogoutHandlerImpl.java
@@ -1,10 +1,9 @@
 package com.renkaen.cat_hospital.service.impl.security;
 
-import com.renkaen.cat_hospital.bean.PO.StaffToken;
+import com.renkaen.cat_hospital.bean.DO.StaffToken;
 import com.renkaen.cat_hospital.utils.RedisUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.security.web.authentication.logout.LogoutHandler;
 import org.springframework.stereotype.Component;
 
@@ -18,7 +17,9 @@ public class LogoutHandlerImpl implements LogoutHandler {
     @Override
     public void logout(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
         System.out.println(authentication);
-        StaffToken staffToken = (StaffToken) authentication.getPrincipal();
-        redisUtil.deleteObject("login:"+staffToken.getUsername());
+        if(authentication!=null) {
+            StaffToken staffToken = (StaffToken) authentication.getPrincipal();
+            redisUtil.deleteObject("login:" + staffToken.getUsername());
+        }
     }
 }
